dashboard/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
flask-jwt-extended>=4.6.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
